Handle tests that call LRO with WaitUntil.Completed
